fix embodied_verse bug
